Update sketch.js
I am responsible for this Character part. Due to the network problem, I failed to push it successfully before, so I sent it directly to the team members for integration. The network problem has been solved. So, I am re-committing this part now. The content of this part remains unchanged.